requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.8
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
urllib3==2.2.2
